+++ src/converters/piral-million/src/extension.ts
@@ -1,0 +1,4 @@
+import type { ExtensionSlotProps } from 'piral-core';
+import { m } from 'million';
+
+export const Extension = (props: ExtensionSlotProps) => m('piral-extension', props);