+++ src/tooling/piral-cli/src/commands.ts
@@ -198,9 +198,12 @@ const allCommands: Array<ToolCommand<any>> = [
         .choices('provider', availableReleaseProviders)
         .describe('provider', 'Sets the provider for publishing the release assets.')
         .default('provider', apps.publishPiralDefaults.provider)
-        .option('fields', undefined)
-        .describe('fields', 'Sets additional fields to be included in the feed service request.')
-        .default('fields', apps.publishPiralDefaults.fields)
+        .option('opts', undefined)
+        .describe('opts', 'Sets the options to forward to the chosen provider.')
+        .default('opts', apps.publishPiralDefaults.opts)
+        .boolean('interactive')
+        .describe('interactive', 'Defines if authorization tokens can be retrieved interactively.')
+        .default('interactive', apps.publishPiralDefaults.interactive)
         .string('base')
         .default('base', process.cwd())
         .describe('base', 'Sets the base directory. By default the current directory is used.');
@@ -211,7 +214,8 @@ const allCommands: Array<ToolCommand<any>> = [
         logLevel: args['log-level'] as LogLevels,
         type: args.type as PiralBuildType,
         provider: args.provider as string,
-        fields: args.fields as Record<string, string>,
+        opts: args.opts as Record<string, string>,
+        interactive: args.interactive as boolean,
       });
     },
   },
@@ -621,6 +625,9 @@ const allCommands: Array<ToolCommand<any>> = [
         .option('headers', undefined)
         .describe('headers', 'Sets additional headers to be included in the feed service request.')
         .default('headers', apps.publishPiletDefaults.headers)
+        .boolean('interactive')
+        .describe('interactive', 'Defines if authorization tokens can be retrieved interactively.')
+        .default('interactive', apps.publishPiletDefaults.interactive)
         .string('base')
         .default('base', process.cwd())
         .describe('base', 'Sets the base directory. By default the current directory is used.');
@@ -639,6 +646,7 @@ const allCommands: Array<ToolCommand<any>> = [
         fields: args.fields as Record<string, string>,
         headers: args.headers as Record<string, string>,
         mode: args.mode as PiletPublishScheme,
+        interactive: args.interactive as boolean,
         _: args,
       });
     },