+++ src/tooling/piral-cli/src/common/browser.ts
@@ -2,14 +2,18 @@ import { log } from './log';
 import { open } from '../external';
 import { config } from './config';
 
+export async function openBrowserAt(address: string) {
+  try {
+    await open(address, undefined);
+  } catch (err) {
+    log('failedToOpenBrowser_0170', err);
+  }
+}
+
 export async function openBrowser(shouldOpen: boolean, port: number, path: string, https?: boolean) {
   if (shouldOpen) {
-    try {
-      const scheme = https ? 'https' : 'http';
-      const address = `${scheme}://${config.host}:${port}${path}`;
-      await open(address, undefined);
-    } catch (err) {
-      log('failedToOpenBrowser_0170', err);
-    }
+    const scheme = https ? 'https' : 'http';
+    const address = `${scheme}://${config.host}:${port}${path}`;
+    await openBrowserAt(address);
   }
 }