+++ src/tooling/piral-cli/src/release.ts
@@ -1,7 +1,18 @@
-import { basename, resolve } from 'path';
-import { copy, fail } from './common';
+import { basename, dirname, relative, resolve } from 'path';
+import { copy, fail, findFile, FormDataObj, postForm, readBinary } from './common';
 import { availableReleaseProviders } from './helpers';
 import { ReleaseProvider } from './types';
+
+async function getVersion(directory: string) {
+  const data = await findFile(directory, 'package.json');
+
+  if (!data) {
+    fail('packageJsonNotFound_0020');
+  }
+
+  const { version } = require(data);
+  return version;
+}
 
 export interface QualifiedReleaseProvider {
   name: string;
@@ -12,7 +23,7 @@ const providers: Record<string, ReleaseProvider> = {
   none() {
     return Promise.resolve();
   },
-  async xcopy(files, args) {
+  async xcopy(_, files, args) {
     const { target } = args;
 
     if (!target) {
@@ -20,6 +31,31 @@ const providers: Record<string, ReleaseProvider> = {
     }
 
     await Promise.all(files.map(async (file) => copy(file, resolve(target, basename(file)))));
+  },
+  async feed(directory, files, args, interactive) {
+    const { url, apiKey, scheme = 'basic', version = await getVersion(directory) } = args;
+
+    if (!url) {
+      fail('publishFeedMissingUrl_0115');
+    }
+
+    if (!version) {
+      fail('publishFeedMissingVersion_0116');
+    }
+
+    const data: FormDataObj = {
+      version,
+      type: 'custom',
+    };
+
+    for (const file of files) {
+      const relPath = relative(directory, file);
+      const fileName = basename(file);
+      const content = await readBinary(dirname(file), fileName);
+      data[relPath] = [content, fileName];
+    }
+
+    await postForm(url, scheme as any, apiKey, data, {}, undefined, interactive);
   },
 };
 
@@ -43,7 +79,13 @@ export function setReleaseProvider(provider: QualifiedReleaseProvider) {
   }
 }
 
-export function publishArtifacts(providerName: string, files: Array<string>, args: Record<string, string>) {
+export function publishArtifacts(
+  providerName: string,
+  directory: string,
+  files: Array<string>,
+  args: Record<string, string>,
+  interactive: boolean,
+) {
   const runRelease = findReleaseProvider(providerName);
-  return runRelease(files, args);
+  return runRelease(directory, files, args, interactive);
 }